added aewa layer to global map
