+++ app/components/maps/PopulationMap.js
@@ -4,10 +4,22 @@ import { getSqlQuery } from 'helpers/map';
 import { BOUNDARY_COLORS } from 'constants/map';
 import BasicMap from './BasicMap';
 
+const SELECTED_AEWA_STYLE = {
+  opacity: 0.5,
+  weight: 0,
+  dashArray: [1, 7],
+  lineCap: 'round',
+  color: 'white',
+  fill: true,
+  fillOpacity: 0.5,
+  fillColor: '#FCF0C5'
+};
+
 class PopulationMap extends BasicMap {
   constructor(props) {
     super(props);
     this.setPopulationColors(props.populations);
+    this.setAewaLayer = this.setAewaLayer.bind(this);
   }
 
   componentDidMount() {
@@ -16,6 +28,13 @@ class PopulationMap extends BasicMap {
     pane.classList.add('-layer-blending');
     this.populationLayerGroup = L.layerGroup();
     this.populationLayerGroup.addTo(this.map);
+  }
+
+  componentDidUpdate(prevProps, prevState) {
+    super.componentDidUpdate(prevProps, prevState);
+    if (prevProps.layers.hasOwnProperty('aewaExtent') && prevProps.layers.aewaExtent !== this.props.layers.aewaExtent) {
+      this.setAewaLayer();
+    }
   }
 
   componentWillReceiveProps(newProps) {
@@ -50,6 +69,37 @@ class PopulationMap extends BasicMap {
         });
       }
     }
+  }
+
+  setAewaLayer() {
+    if (!this.selectedAewaLayer) {
+      const query = `
+         SELECT ST_AsGeoJSON(the_geom, 15, 1) as geom
+         FROM aewa_extent_geo LIMIT 1
+       `; // asGeoJSON with options - add bbox for fitBound
+      getSqlQuery(query)
+        .then(this.addAewaLayer.bind(this));
+    } else {
+      this.selectedAewaLayer.remove(this.map);
+      this.selectedAewaLayer = null;
+    }
+  }
+
+  addAewaLayer(data) {
+    // layer not found, just set map view on selectedSite with default zoom
+    if (!data.rows.length) {
+      this.map.setView([this.props.selectedSite.lat, this.props.selectedSite.lon], 8);
+      return;
+    }
+
+    const geom = JSON.parse(data.rows[0].geom);
+    const layer = L.geoJSON(geom, {
+      noWrap: true,
+      style: SELECTED_AEWA_STYLE
+    });
+    layer.addTo(this.map);
+    layer.bringToBack();
+    this.selectedAewaLayer = layer;
   }
 
   setPopulationColors(populations) {
